+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 * ### NI-DCPower
   * #### Added
   * #### Changed
+    * Metadata updated to NI-DCPower 17.6.1
   * #### Removed
 * ### NI-FGEN
   * #### Added
@@ -237,3 +238,4 @@ and this project adheres to [Python Versioning](http://legacy.python.org/dev/pep
   * #### Changed
   * #### Removed
 -->
+